requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.10.0
cffi==1.17.1
click==8.1.8
contextlib2==21.6.0
cryptography==44.0.0
Django==5.1.4
django-cors-headers==4.6.0
django-ninja==1.3.0
django-ninja-extra==0.22.0
django-ninja-jwt==5.3.5
injector==0.22.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
